backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML Dependencies
openai==1.3.0
xgboost==2.0.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# API Integration
httpx==0.25.1
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS Integration
boto3==1.29.3
botocore==1.32.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1 